refactor(dashboard): resolve React Doctor warnings in settings
