+++ apps/dashboard/src/components/integrations/github/github-repository-actions.tsx
@@ -1,169 +1,36 @@
 "use client";
 
-import { MoreHorizontalIcon } from "@hugeicons/core-free-icons";
-import { HugeiconsIcon } from "@hugeicons/react";
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuTrigger,
-} from "@notra/ui/components/ui/dropdown-menu";
-import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
-import { useState } from "react";
-import { toast } from "sonner";
-
-import { Button } from "@/components/button";
 import { DeleteIntegrationDialog } from "@/components/delete-integration-dialog";
 import { EditIntegrationDialog } from "@/components/integrations/edit-integration-dialog";
+import { GitHubRepositoryMenu } from "@/components/integrations/github/github-repository-menu";
 import { LegacyEditTokenDialog } from "@/components/integrations/legacy/edit-token-dialog";
-import { dashboardOrpc } from "@/lib/orpc/query";
-import type {
-  GitHubRepositoryActionsProps,
-  GitHubRepositoryDialog,
-} from "@/types/integrations/github";
+import { useGitHubRepositoryActions } from "@/hooks/use-github-repository-actions";
+import type { GitHubRepositoryActionsProps } from "@/types/integrations/github";
 
-export function GitHubRepositoryActions({
-  onMigrate,
-  isMigrating,
-  onToggleWebhooks,
-  webhooksOpen,
-  integration,
-  organizationId,
-  onManageRepositories,
-}: GitHubRepositoryActionsProps) {
-  const queryClient = useQueryClient();
-  const isEnabled =
-    integration.enabled &&
-    integration.repositories.every((repository) => repository.enabled);
-  const [dialog, setDialog] = useState<GitHubRepositoryDialog>(null);
-  const affectedSchedules = useQuery({
-    ...dashboardOrpc.integrations.affectedSchedules.queryOptions({
-      input: { organizationId, integrationId: integration.id },
-    }),
-    enabled: dialog === "delete",
-  });
-  const invalidate = () =>
-    Promise.all([
-      queryClient.invalidateQueries({
-        queryKey: dashboardOrpc.integrations.key(),
-      }),
-      queryClient.invalidateQueries({
-        queryKey: dashboardOrpc.github.app.get.queryKey({
-          input: { organizationId },
-        }),
-      }),
-    ]);
-  const toggle = useMutation({
-    mutationFn: async () => {
-      if (!isEnabled) {
-        await Promise.all(
-          integration.repositories
-            .filter((repository) => !repository.enabled)
-            .map((repository) =>
-              dashboardOrpc.integrations.repositories.update.call({
-                organizationId,
-                repositoryId: repository.id,
-                enabled: true,
-              })
-            )
-        );
-      }
-      return dashboardOrpc.integrations.update.call({
-        organizationId,
-        integrationId: integration.id,
-        enabled: !isEnabled,
-      });
-    },
-    onSuccess: async () => {
-      await invalidate();
-      toast.success(isEnabled ? "Repository paused" : "Repository enabled");
-    },
-    onError: (error) => toast.error(error.message),
-  });
-  const remove = useMutation({
-    mutationFn: () =>
-      dashboardOrpc.integrations.delete.call({
-        organizationId,
-        integrationId: integration.id,
-      }),
-    onSuccess: async () => {
-      await Promise.all([
-        invalidate(),
-        queryClient.invalidateQueries({
-          queryKey: dashboardOrpc.automation.key(),
-        }),
-      ]);
+export function GitHubRepositoryActions(props: GitHubRepositoryActionsProps) {
+  const { integration, organizationId } = props;
+  const { isEnabled, dialog, setDialog, affectedSchedules, toggle, remove } =
+    useGitHubRepositoryActions(props);
+  const handleOpenChange = (open: boolean) => {
+    if (!open) {
       setDialog(null);
-      toast.success("Repository removed");
-    },
-    onError: (error) => toast.error(error.message),
-  });
-
+    }
+  };
   return (
     <>
-      <DropdownMenu>
-        <DropdownMenuTrigger
-          render={
-            <Button
-              aria-label={`Manage ${integration.displayName}`}
-              disabled={toggle.isPending || remove.isPending}
-              size="icon-sm"
-              variant="ghost"
-            />
-          }
-        >
-          <HugeiconsIcon icon={MoreHorizontalIcon} className="size-4" />
-        </DropdownMenuTrigger>
-        <DropdownMenuContent align="end">
-          {!integration.managedByGitHubApp ? (
-            <>
-              <DropdownMenuItem onClick={() => setDialog("edit")}>
-                Edit repository
-              </DropdownMenuItem>
-              <DropdownMenuItem onClick={() => setDialog("token")}>
-                Update access token
-              </DropdownMenuItem>
-            </>
-          ) : null}
-          {integration.managedByGitHubApp ? (
-            <DropdownMenuItem onClick={onManageRepositories}>
-              Manage selection
-            </DropdownMenuItem>
-          ) : (
-            <DropdownMenuItem onClick={() => toggle.mutate()}>
-              {isEnabled ? "Pause repository" : "Enable repository"}
-            </DropdownMenuItem>
-          )}
-          {!integration.managedByGitHubApp ? (
-            <>
-              <DropdownMenuItem disabled={isMigrating} onClick={onMigrate}>
-                {isMigrating ? "Switching…" : "Switch to GitHub App"}
-              </DropdownMenuItem>
-              {integration.repositories.length > 0 ? (
-                <DropdownMenuItem onClick={onToggleWebhooks}>
-                  {webhooksOpen ? "Hide webhooks" : "Webhook settings"}
-                </DropdownMenuItem>
-              ) : null}
-            </>
-          ) : null}
-          <DropdownMenuItem
-            onClick={() => setDialog("delete")}
-            variant="destructive"
-          >
-            Remove repository
-          </DropdownMenuItem>
-        </DropdownMenuContent>
-      </DropdownMenu>
+      <GitHubRepositoryMenu
+        {...props}
+        isEnabled={isEnabled}
+        isPending={toggle.isPending || remove.isPending}
+        onToggle={() => toggle.mutate()}
+        onDialog={setDialog}
+      />
       {dialog === "edit" ? (
         <EditIntegrationDialog
           integration={integration}
           organizationId={organizationId}
           open
-          onOpenChange={(open) => {
-            if (!open) {
-              setDialog(null);
-            }
-          }}
+          onOpenChange={handleOpenChange}
         />
       ) : null}
       {dialog === "token" ? (
@@ -171,11 +38,7 @@ export function GitHubRepositoryActions({
           integration={integration}
           organizationId={organizationId}
           open
-          onOpenChange={(open) => {
-            if (!open) {
-              setDialog(null);
-            }
-          }}
+          onOpenChange={handleOpenChange}
         />
       ) : null}
       <DeleteIntegrationDialog
@@ -185,11 +48,7 @@ export function GitHubRepositoryActions({
         isLoadingSchedules={affectedSchedules.isLoading}
         onConfirm={() => remove.mutate()}
         open={dialog === "delete"}
-        onOpenChange={(open) => {
-          if (!open) {
-            setDialog(null);
-          }
-        }}
+        onOpenChange={handleOpenChange}
       />
     </>
   );
