+++ apps/dashboard/src/components/settings/panes/logs-pane.tsx
@@ -1,26 +1,5 @@
 "use client";
 
-import {
-  Cancel01Icon,
-  InformationCircleIcon,
-  Search01Icon,
-} from "@hugeicons/core-free-icons";
-import { HugeiconsIcon } from "@hugeicons/react";
-import { FEATURES } from "@notra/ai/billing/features";
-import { Input } from "@notra/ui/components/ui/input";
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from "@notra/ui/components/ui/select";
-import {
-  Tooltip,
-  TooltipContent,
-  TooltipTrigger,
-} from "@notra/ui/components/ui/tooltip";
-import { useDebouncedCallback } from "@tanstack/react-pacer";
 import { useQuery } from "@tanstack/react-query";
 import {
   parseAsInteger,
@@ -35,49 +14,24 @@ import { DataTable } from "@/app/(dashboard)/[slug]/settings/logs/data-table";
 import { LogsPageSkeleton } from "@/app/(dashboard)/[slug]/settings/logs/skeleton";
 import { Button } from "@/components/button";
 import { LogDetailsSheet } from "@/components/logs/log-details-sheet";
+import { LogFilters } from "@/components/logs/log-filters";
+import { LogRetentionHint } from "@/components/logs/log-retention-hint";
 import type { SortState } from "@/components/motion/table/types";
 import { useOrganizationsContext } from "@/components/providers/organization-provider";
 import { SettingsPane } from "@/components/settings/settings-pane";
 import {
-  SOURCE_LABELS,
   SOURCE_VALUES,
-  STATUS_LABELS,
   STATUS_VALUES,
   LOGS_PAGE_SIZE,
   LOGS_OVERVIEW_STALE_TIME_MS,
 } from "@/constants/logs";
-import { useBillingCustomer } from "@/lib/hooks/use-billing-customer";
 import { dashboardOrpc } from "@/lib/orpc/query";
 import type { LogSelection } from "@/types/logs/details-sheet";
-import type { LogsRetentionHintProps } from "@/types/settings/modal";
-import type { Log } from "@/types/webhooks/webhooks";
 import { getLogPage } from "@/utils/log-pagination";
-import { getSourceLabel, getStatusLabel } from "@/utils/logs";
-
-const SEARCH_DEBOUNCE_MS = 150;
-
-function LogsRetentionHint({ days }: LogsRetentionHintProps) {
-  return (
-    <Tooltip>
-      <TooltipTrigger
-        aria-label="Log retention information"
-        className="text-muted-foreground hover:text-foreground inline-flex cursor-help transition-colors"
-      >
-        <HugeiconsIcon className="size-3.5" icon={InformationCircleIcon} />
-      </TooltipTrigger>
-      <TooltipContent>
-        Log data is retained for {days} days. Older entries are automatically
-        removed.
-      </TooltipContent>
-    </Tooltip>
-  );
-}
 
 export function LogsSettingsPane() {
   const { activeOrganization } = useOrganizationsContext();
-  const organizationId = activeOrganization?.id;
-  const { check, data: customer } = useBillingCustomer();
-
+  const organizationId = activeOrganization?.id ?? "";
   const [page, setPage] = useQueryState("page", parseAsInteger.withDefault(1));
   const [search, setSearch] = useQueryState("q", parseAsString.withDefault(""));
   const [source, setSource] = useQueryState(
@@ -88,62 +42,21 @@ export function LogsSettingsPane() {
     "status",
     parseAsStringLiteral(STATUS_VALUES).withDefault("all")
   );
-  const [searchInput, setSearchInput] = useState(search);
-  const [prevSearch, setPrevSearch] = useState(search);
   const [selection, setSelection] = useState<LogSelection | null>(null);
   const selectedLog =
-    selection && selection.organizationId === organizationId
-      ? selection.log
-      : null;
+    selection?.organizationId === organizationId ? selection.log : null;
   const [sort, setSort] = useState<SortState | null>({
     key: "createdAt",
     direction: "desc",
   });
-
-  if (search !== prevSearch) {
-    setPrevSearch(search);
-    setSearchInput(search);
-  }
-
-  const commitSearch = useDebouncedCallback(
-    (value: string) => {
-      setSearch(value);
-      if (page !== 1) {
-        setPage(1);
-      }
-    },
-    { wait: SEARCH_DEBOUNCE_MS }
-  );
-
-  const has30DayRetention = customer
-    ? check({ featureId: FEATURES.LOG_RETENTION_30_DAYS }).allowed
-    : false;
-  const has14DayRetention = customer
-    ? check({ featureId: FEATURES.LOG_RETENTION_14_DAYS }).allowed
-    : false;
-  let logRetentionDays = 7;
-  if (has30DayRetention) {
-    logRetentionDays = 30;
-  } else if (has14DayRetention) {
-    logRetentionDays = 14;
-  }
-
-  const resetPage = () => {
-    if (page !== 1) {
-      setPage(1);
-    }
-  };
-
+  const resetPage = () => setPage(1);
   const logsQuery = useQuery({
     ...dashboardOrpc.logs.webhooks.overview.queryOptions({
-      input: {
-        organizationId: organizationId ?? "",
-      },
+      input: { organizationId },
     }),
-    enabled: !!organizationId,
+    enabled: Boolean(organizationId),
     staleTime: LOGS_OVERVIEW_STALE_TIME_MS,
   });
-
   const result = getLogPage(logsQuery.data?.logs ?? [], {
     page,
     pageSize: LOGS_PAGE_SIZE,
@@ -152,108 +65,36 @@ export function LogsSettingsPane() {
     search,
     sort,
   });
-
   const filtersActive =
     source !== "all" || status !== "all" || search.length > 0;
-
   const resetFilters = () => {
-    setSearchInput("");
     setSearch("");
     setSource("all");
     setStatus("all");
     resetPage();
   };
-
   return (
-    <SettingsPane
-      titleAccessory={<LogsRetentionHint days={logRetentionDays} />}
-    >
-      <div className="flex flex-col gap-2 sm:flex-row sm:items-center">
-        <div className="relative flex-1">
-          <HugeiconsIcon
-            className="text-muted-foreground absolute top-1/2 left-2.5 size-4 -translate-y-1/2"
-            icon={Search01Icon}
-          />
-          <Input
-            aria-label="Search logs"
-            autoComplete="off"
-            className="pr-8 pl-8"
-            name="log-search"
-            onChange={(e) => {
-              setSearchInput(e.target.value);
-              commitSearch(e.target.value);
-            }}
-            placeholder="Search by title or error message"
-            type="text"
-            value={searchInput}
-          />
-          {searchInput.length > 0 && (
-            <button
-              aria-label="Clear search"
-              className="text-muted-foreground hover:bg-muted hover:text-foreground focus-visible:ring-ring absolute top-1/2 right-1 flex size-6 -translate-y-1/2 items-center justify-center rounded-md transition-colors focus-visible:ring-2 focus-visible:outline-none"
-              onClick={() => {
-                setSearchInput("");
-                commitSearch("");
-              }}
-              type="button"
-            >
-              <HugeiconsIcon
-                aria-hidden="true"
-                className="size-3.5"
-                icon={Cancel01Icon}
-              />
-            </button>
-          )}
-        </div>
-        <Select
-          onValueChange={(value) => {
-            setSource(value);
-            resetPage();
-          }}
-          value={source}
-        >
-          <SelectTrigger aria-label="Filter by source" className="sm:w-44">
-            <SelectValue>
-              {(value: string) => getSourceLabel(value)}
-            </SelectValue>
-          </SelectTrigger>
-          <SelectContent>
-            {SOURCE_VALUES.map((value) => (
-              <SelectItem key={value} value={value}>
-                {SOURCE_LABELS[value]}
-              </SelectItem>
-            ))}
-          </SelectContent>
-        </Select>
-        <Select
-          onValueChange={(value) => {
-            setStatus(value);
-            resetPage();
-          }}
-          value={status}
-        >
-          <SelectTrigger aria-label="Filter by status" className="sm:w-40">
-            <SelectValue>
-              {(value: string) => getStatusLabel(value)}
-            </SelectValue>
-          </SelectTrigger>
-          <SelectContent>
-            {STATUS_VALUES.map((value) => (
-              <SelectItem key={value} value={value}>
-                {STATUS_LABELS[value]}
-              </SelectItem>
-            ))}
-          </SelectContent>
-        </Select>
-        <Button
-          variant="outline"
-          disabled={logsQuery.isFetching}
-          onClick={() => logsQuery.refetch()}
-        >
-          {logsQuery.isFetching && logsQuery.data ? "Refreshing…" : "Refresh"}
-        </Button>
-      </div>
-
+    <SettingsPane titleAccessory={<LogRetentionHint />}>
+      <LogFilters
+        search={search}
+        source={source}
+        status={status}
+        onSearchChange={(value) => {
+          setSearch(value);
+          resetPage();
+        }}
+        onSourceChange={(value) => {
+          setSource(value);
+          resetPage();
+        }}
+        onStatusChange={(value) => {
+          setStatus(value);
+          resetPage();
+        }}
+        onRefresh={() => logsQuery.refetch()}
+        isFetching={logsQuery.isFetching}
+        hasData={Boolean(logsQuery.data)}
+      />
       {logsQuery.isError ? (
         <div
           role="alert"
@@ -295,11 +136,7 @@ export function LogsSettingsPane() {
                 }
           }
           onPageChange={setPage}
-          onRowClick={(log: Log) => {
-            if (organizationId) {
-              setSelection({ organizationId, log });
-            }
-          }}
+          onRowClick={(log) => setSelection({ organizationId, log })}
           page={result.page}
           totalPages={result.totalPages}
           totalCount={result.totalCount}
@@ -307,7 +144,7 @@ export function LogsSettingsPane() {
       ) : null}
       <LogDetailsSheet
         key={organizationId}
-        organizationId={organizationId ?? ""}
+        organizationId={organizationId}
         organizationSlug={activeOrganization?.slug ?? ""}
         log={selectedLog}
         onOpenChange={(open) => {
